Move over entries APIs
