+++ src/api/entries.ts
@@ -1,11 +1,27 @@
-import { Context } from 'hono'
+import { ContextWithBindings } from '../common/bindings'
+import { parseConfigEntry } from '../common/parse'
+import { errorResponse, successResponse } from '../common/response'
+import { ConfigEntry, ConfigValue } from '../common/types'
 
-export const entriesGet = (c: Context) => {
-  console.log('Handling GET request for entries')
-  return c.json({ entries: [] })
-}
-
-export const entriesPost = (c: Context) => {
+export const entriesPost = async (c: ContextWithBindings) => {
   console.log('Handling POST request for entries')
-  return c.json({ message: 'success' })
+
+  let entry: ConfigEntry
+  try {
+    entry = parseConfigEntry(await c.req.text())
+  } catch (e) {
+    const error = e as Error
+    return errorResponse(c, error)
+  }
+
+  const value = {
+    type: entry.type,
+    value: entry.value
+  } as ConfigValue
+
+  console.log(`Storing key entry:${entry.key} in KV`)
+  await c.env.CONFIG_KV.put(`entry:${entry.key}`, JSON.stringify(value))
+  console.log(`Stored key entry:${entry.key} in KV`)
+
+  return successResponse(c)
 }
